requirements.txt
discord.py==1.5.1
python-dotenv==0.15.0
peewee==3.14.0
pymysql==0.10.1
discord_pretty_help==1.2.0
flask==1.1.2
flask_admin==1.5.7
flask_login==0.5.0
Werkzeug==1.0.1
WTForms==2.3.3
wtf-peewee==3.0.2
email-validator==1.1.2
